+++ src/main/java/com/prospect/crm/repository/RolePermissionRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RolePermissionRepository extends JpaRepository<RolePermission, Integer> {
+public interface RolePermissionRepository extends JpaRepository<RolePermission, Long> {
     List<RolePermission> findByRoleId(Role roleId);
     List<RolePermission> findByPermission(String permission);
 }
